+++ chaosgcp/storage/__init__.py
@@ -1,0 +1,18 @@
+# -*- coding: utf-8 -*-
+
+from chaoslib.types import Configuration, Secrets
+from google.cloud import storage as gc_storage
+
+from chaosgcp import load_credentials, get_context
+
+__all__ = ["client"]
+
+
+def client(configuration: Configuration = None,
+           secrets: Secrets = None):
+    """
+    Create a client for Google Cloud Storage
+    """
+    ctx = get_context(configuration=configuration, secrets=secrets)
+    credentials = load_credentials(secrets=secrets)
+    return gc_storage.Client(project=ctx.project_id, credentials=credentials)